tests/requirements.txt
parameterized
backports.lzma  ;  python_version < '3.3'
